Move delete_comment and submit_report into blueprint
