+++ nyaa/views/torrents.py
@@ -174,6 +174,59 @@ def download_torrent(torrent_id):
     return resp
 
 
+@bp.route('/view/<int:torrent_id>/comment/<int:comment_id>/delete', methods=['POST'])
+def delete_comment(torrent_id, comment_id):
+    if not flask.g.user:
+        flask.abort(403)
+    torrent = models.Torrent.by_id(torrent_id)
+    if not torrent:
+        flask.abort(404)
+
+    comment = models.Comment.query.filter_by(id=comment_id).first()
+    if not comment:
+        flask.abort(404)
+
+    if not (comment.user.id == flask.g.user.id or flask.g.user.is_moderator):
+        flask.abort(403)
+
+    db.session.delete(comment)
+    db.session.flush()
+    torrent.update_comment_count()
+
+    url = flask.url_for('torrents.view', torrent_id=torrent.id)
+    if flask.g.user.is_moderator:
+        log = "Comment deleted on torrent [#{}]({})".format(torrent.id, url)
+        adminlog = models.AdminLog(log=log, admin_id=flask.g.user.id)
+        db.session.add(adminlog)
+    db.session.commit()
+
+    flask.flash('Comment successfully deleted.', 'success')
+
+    return flask.redirect(url)
+
+
+@bp.route('/view/<int:torrent_id>/submit_report', endpoint='report', methods=['POST'])
+def submit_report(torrent_id):
+    if not flask.g.user:
+        flask.abort(403)
+
+    form = forms.ReportForm(flask.request.form)
+
+    if flask.request.method == 'POST' and form.validate():
+        report_reason = form.reason.data
+        current_user_id = flask.g.user.id
+        report = models.Report(
+            torrent_id=torrent_id,
+            user_id=current_user_id,
+            reason=report_reason)
+
+        db.session.add(report)
+        db.session.commit()
+        flask.flash('Successfully reported torrent!', 'success')
+
+    return flask.redirect(flask.url_for('torrents.view', torrent_id=torrent_id))
+
+
 @bp.route('/upload', methods=['GET', 'POST'])
 def upload():
     upload_form = forms.UploadForm(CombinedMultiDict((flask.request.files, flask.request.form)))
